requirements.txt
numpy
scipy
numpy-quaternion
matplotlib
cv2 
gym